chore: refactor fuel-management page
